Updated doco to reflect new Lambda-based access control policies (#13) [skip ci]
A built-in Javascript VM is both a) more complex to implement and b) less
flexible for users. AWS already has a service for this exact kind of glue logic
in the form of Lambda. What's more, this pattern of factoring out authorisation
already exists in the form of API Gateway Custom Authorizers, so users are
probably already familiar/comfortable with it.