Allow new KMS key id format (relates to #29)
The KMS Encrypt API call accepts either a key ID, an alias or a full ARN of
either of those. What it doesn't support is an <account-id>:key/<key-id> syntax
like Lambda does. Allowing users to specify a key in this format is desirable
in the common enterprise scenario where KMS keys are hosted in a different
account and the same alias can be used in different regions. Now consumers can
pass --kms-key <account-id>:alias/MyAlias in userdata (or baked into an AMI) and
it will work in any region.